+++ docs/ChangeLog.md
@@ -50,7 +50,7 @@
 - make: work around `make-3.81` bug of pattern rules `#D2065` f7ec170b
 - decode: work around `convert-meta on` in bash >= 5.2 with broken locale (reported by 3ximus) `#D2069` 226f9718
 - canvas: adjust GraphemeClusterBreak of hankaku-kana voiced marks `#D2077` 31d168cc
-- canvas: update tables and grapheme clusters for Unicode 15.1.0 `#D2078` xxxxxxxx xxxxxxxx
+- canvas: update tables and grapheme clusters for Unicode 15.1.0 `#D2078` 503bb38b 9d84b424 xxxxxxxx
 
 ## Contrib